backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
django-celery-beat==2.7.0
django-storages==1.14.4
gunicorn==23.0.0
psycopg2-binary==2.9.10
celery[redis]==5.4.0
redis==5.2.1
langchain==0.3.20
langchain-openai==0.3.9
langchain-community==0.3.12
langchain-chroma==0.2.2
langgraph==0.3.21
chromadb==0.5.23
openai==1.66.3
tiktoken==0.8.0
pypdf==5.4.0
python-dotenv==1.0.1
drf-spectacular==0.28.0
Pillow==11.0.0
reportlab==4.2.5
whitenoise==6.8.2
channels==4.2.0
channels-redis==4.2.1
